% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_cubeinfo.R
\name{aux_cubeinfo}
\alias{aux_cubeinfo}
\title{Get cube file information}
\usage{
aux_cubeinfo(file, gipptools)
}
\arguments{
\item{file}{\code{Characater} value, cube file to be processes}

\item{gipptools}{\code{Character} value, path to gipptools or 
Gipptools directory.}
}
\value{
\code{data frame} with cube meta data
}
\description{
This is a simple wrapper for the Gipptools program \code{cubeinfo}, 
providing a short summary of the cube file meta data, in a coherent 
data frame structure.
}
\examples{

\dontrun{

## get cube info
x = aux_cubeinfo(file = "data/cube/example.ATB", 
                 gipptools = "/software/gipptools-2019.332/")

}

}
\author{
Michael Dietze
}
\keyword{eseis}
