% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_getFDSNdata.R
\name{aux_getFDSNdata}
\alias{aux_getFDSNdata}
\title{Download seismic data from FDSN data base}
\usage{
aux_getFDSNdata(
  start,
  duration,
  component,
  network,
  station,
  url,
  link_only = FALSE,
  eseis = TRUE
)
}
\arguments{
\item{start}{\code{POSIXct} value, start time of the data to query.}

\item{duration}{\code{Numeric} value, length of the data to query, in 
seconds.}

\item{component}{\code{Character} value, seismic component or channel to 
get.}

\item{network}{\code{Character} value, two-character FDSN network ID.}

\item{station}{\code{Character} value, FDSN station ID.}

\item{url}{\code{Character} value, FDSN URL. See details for further 
information.}

\item{link_only}{\code{Logical} value, return only FDSN link instead of
downloading and importing the data.}

\item{eseis}{\code{Logical} scalar, option to read data to an \code{eseis}
object (recommended, see documentation of 
\code{aux_initiateeseis}), default is \code{TRUE}}
}
\value{
\code{List} object with imported seismic data for each provided 
set of input arguments.
}
\description{
The function accesses the specified FDSN internet data base and 
downloads seismic data based on the network and station IDs and time
constraints.
}
\details{
A convenient way to get all the required input data is using the 
function \code{aux_getFDSNstation} before. It will return all the 
information in a structured way.

The FDSN web service URL must be provided only with its base path, e.g.
\code{"http://service.iris.edu"}.
}
\examples{

\dontrun{

## get stations < 0.6 degrees away from Piz Chengalo collapse
x <- aux_getFDSNstation(centre = c(46.3, 9.6),
                        radius = 0.6,
                        access = TRUE)

## sort statiions by distance
x <- x[order(x$distance),]

## download available data
d <- aux_getFDSNdata(start = "2017-08-23 07:30:00",
                     duration = 180, 
                     component = "BHZ",
                     network = x$network, 
                     station = x$station, 
                     url = x$url)

## remove stations without available data
x <- x[!unlist(lapply(d, is.null)),]
d <- d[!unlist(lapply(d, is.null))]

## generate plots of the three nearest stations
par(mfcol = c(3, 1))

for(i in 1:3) {

  plot_signal(data = d[[i]],
              main = paste(x$ID[i], 
                           " | ",
                           round(x$distance[i], 2),
                           "distance (DD)"))
} 
}
                     
}
\seealso{
aux_get_FDSNstation, read_mseed
}
\author{
Michael Dietze
}
\keyword{eseis}
