% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_initiateeseis.R
\name{aux_initiateeseis}
\alias{aux_initiateeseis}
\title{Initiate an eseis object}
\usage{
aux_initiateeseis()
}
\value{
\code{S3} list object of class \code{eseis}.
}
\description{
The function generates an empty eseis object, starting with processing 
step 0. The object contains no data and the history only contains the 
system information.
}
\examples{

## initiate eseis object
aux_initiateeseis()
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
