% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_organisefdsnfiles.R
\name{aux_organisefdsnfiles}
\alias{aux_organisefdsnfiles}
\title{Download and organise FDSN data sets}
\usage{
aux_organisefdsnfiles(
  start,
  stop,
  network,
  station,
  component,
  url,
  dir,
  format = "mseed",
  pattern = "eseis",
  buffer = c(300, 300),
  verbose = FALSE
)
}
\arguments{
\item{start}{\code{POSIXct} value, start time for data acquisition. If a 
character string (or vector) is provided, the function will try to convert 
it to POSIXct. Note that the start time will be rounded down to the full
hour.}

\item{stop}{\code{POSIXct} value, stop time for data acquisition. If a 
character string (or vector) is provided, the function will try to convert 
it to POSIXct. Note that the stop time will be rounded up to the full
hour.}

\item{network}{\code{Character} vector, two-character FDSN network ID.}

\item{station}{\code{Character} vector, FDSN station ID.}

\item{component}{\code{Character} vector, seismic channel or channels to 
get. Default is \code{"BHZ"}. Note that the component(s) must exist in the 
FDSN data repository. See details for further information.}

\item{url}{\code{Character} vector, FDSN URL.}

\item{dir}{\code{Character} value, path to directory where downloaded 
data is written to.}

\item{format}{\code{Character} value, output file format. One out of 
\code{"mseed"} and \code{"sac"}. Default is \code{"mseed"}.}

\item{pattern}{\code{Character} value, file organisation scheme keyword. 
One out of \code{"eseis"} and \code{"seiscomp"}. Default is \code{"eseis"}. 
See details and \code{read_data} for further information.}

\item{buffer}{\code{Numeric} vector of length two, size of the two-sided 
buffer around the data to download, in seconds. Default is 
\code{c(300, 300)}. The buffer is useful because some data do not start 
at the desired start and end times. The buffer is only used for downloading
the data and will be clipped before writing the files.}

\item{verbose}{\code{Logical} value, option to enable informative screen 
output. Default is \code{FALSE}.}
}
\value{
Set of seismic files organised for seamless use by eseis functions
}
\description{
This function downloads and organises seismic files provided by FDSN 
data services.
}
\author{
Michael Dietze
}
\keyword{eseis}
