% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{package}
\name{eseis}
\alias{eseis}
\alias{eseis-package}
\title{eseis: Environmental Seismology Toolbox}
\description{
The package eseis contains a growing set of function to handle the 
complete workflow of environmental seismology, i.e., the scientific 
field that studies the seismic signals that are emitted by Earth 
surface processes. The package suppers reading the two most common 
seismic data formats, general functions for preparational and 
analytical signal processing as well as specified functions for 
handling signals generated by Earth surface processes. Finally, 
graphical plot functions are provided, too.
}
\author{
Michael Dietze
}
\keyword{Environmental}
\keyword{Seismology,}
\keyword{processing}
\keyword{signal}
