% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_models.R
\name{gui_models}
\alias{gui_models}
\title{Start GUI with seismic models}
\usage{
gui_models(...)
}
\arguments{
\item{...}{further arguments to pass to \code{runApp}}
}
\description{
This function starts a browser-based graphic user interface to explore the 
parameter space of seismic models that predict the spectra of turbulent 
water flow and bedload flux. Empirical spectra can be added if they are 
present in the current environment as eseis objects. Note that those 
spectra should be made from deconvolved input data in order to have the 
same units as the model spectra (dB). ATTENTION, due to an unresolved 
issue, there must be at least one eseis object present in the working 
environment.
}
\examples{

\dontrun{
# Start the GUI
gui_models()
}

}
\seealso{
\code{runApp}
}
\author{
Michael Dietze
}
