% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_logger.R
\name{list_logger}
\alias{list_logger}
\title{List logger library.}
\usage{
list_logger()
}
\value{
\code{List} object, supported loggers with their parameters.
}
\description{
The function returns the list of supported data loggers to extract signal 
deconvolution parameters.
}
\examples{

## show sensors
list_logger
                     
}
\author{
Michael Dietze
}
\keyword{eseis}

