% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rainfall.R
\name{model_rainfall}
\alias{model_rainfall}
\title{Model the seismic spectrum due to rainfall}
\usage{
model_rainfall(
  I,
  depth = 0,
  q_0,
  v_0,
  e_0,
  p_0,
  n_0 = 1,
  r_s = 2650,
  f = c(1, 100),
  res = 1000,
  eseis = TRUE,
  ...
)
}
\arguments{
\item{I}{\code{Numeric} value, rain intensity (mm/h)}

\item{depth}{\code{Numeric} value, sensor depth below surface (m),
if omitted, default value is \code{0}.}

\item{q_0}{\code{Numeric} value, ground quality factor at \code{f_0}}

\item{v_0}{\code{Numeric} value, phase velocity at \code{f_0} (m/s)}

\item{e_0}{\code{Numeric} value, variation exponent of quality factor change
with frequency.}

\item{p_0}{\code{Numeric} value, variation exponent of Rayleigh wave 
velocity with frequency}

\item{n_0}{\code{Numeric} value, Greens function amplitude scaling 
coefficient, assumed to be near \code{1} at high frequencies (Tsai and 
Atiganyanun, 2014). Default is \code{1}.}

\item{r_s}{\code{Numeric} value, specific density of the ground, default 
is \code{2650} kg/m^3}

\item{f}{\code{Numeric} vector, frequency range to be modelled. If of 
length two the argument is interpreted as representing the lower and 
upper limit and the final length of the frequency vector is set by the 
argument \code{res}. If f contains more than two values it is interpreted 
as the actual frequency vector and the value of \code{res} is ignored. 
Default is \code{c(1, 100)}.}

\item{res}{\code{Numeric} value, output resolution, i.e. length of the 
spectrum vector. Default is \code{1000}.}

\item{eseis}{\code{Logical} value, option to return an eseis object 
instead of a data frame. Default is \code{FALSE}.}

\item{\dots}{Further arguments passed to the function.}
}
\value{
\code{eseis} object containing the modelled spectrum.
}
\description{
The function calculates the seismic spectrum as predicted by the model of 
Porter et al. (2024). The code is adopted from a Matlab script by R. Porter
and integrated to the R package 'eseis' by Michael Dietze. The function 
calculates a rain drop size distribution with the Gamma function approach 
by Uijlenhoet and Stricker (1999). Alternatively, the user can provide an 
empirical drop size distribution and corresponing fall velocities (see 
details). From the input data, the function derives an impulse flux. The 
integrated impulse flux is corrected by the Greens function and returned 
as the seismic power spectrum (\code{eseis} object by default).
}
\details{
The model uses by default the approach by Uijlenhoet and Stricker (1999) to 
convert the bulk rain intensity (\code{I}) to a rain drop size and velocity 
distribution. This requires a vector of drop size limits (\code{D_size}), 
which is by default calculated automatically as a linear spaced sequence of 
values between the limits (\code{D_lim = c(0.1, 10)} (mm)) with a resolution
of 100 values (\code{D_res = 100}). The approach further uses the empirical 
values of Uijlenhoet and Stricker (1999), i.e. \code{a = 3.778} and
\code{B = 0.67}. See list below for details and note that those four  
parameters can also be changed by the user. Then, the  following steps are 
implemented specifically: (1) calculate a factor Gamma following the Gamma 
function with the empirical fit parameter B 
(\code{gamma <- pracma::gammaz(z = 1 + B)}), (2) calculate a factor Delta 
from the rain intensity (\code{delta <- 4.23*I^(-0.214)}), (3) derive a drop 
size distribution function 
(\code{D_prob <- (delta^(1 + B) / gamma) * (D_size^B * exp(-delta * D_size))}), 
(4) estimate the fall velocity for each drop size class 
(\code{D_velo <- a * D_size^B}) ans set velocities above a pre-defined 
maximum fall velocity (\code{v_max = 9.5} m/s) to that maximum value.

Alternatively, a user can also provide empirical information on these rain
parameters. If the following vectors are provided, the 
function will operate with these values instead of estimating them following
the above approach. Note that all vectors must have the same length and that 
no fall velocity correction according to \code{v_max} is performed in that 
case.

(1) \code{D_size}, upper drop size limits (mm)

(2) \code{D_prob}, relative probability of that size class

(3) \code{D_velo}, fall velocity of the corresponding size class (m/s)

The model uses a set of further, predefined constants. These can be changed  
by the user, using the \code{...}-argument:
\itemize{
  \item \code{D_lim = c(0.01, 10)}, range of drop diameters (mm),
  defined as min and max The drop sizes will be linearly interpolated 
  in that range with equal intervals using by default a resolution of 100 
  values (see \code{D_res})
  \item \code{D_res = 100}, resolution of the drop diameter vector defined 
  by the range parameter \code{D}
  \item \code{a = 3.778}, fit parameter for drop velocity from diameters, 
  see Uijlenhoet and Stricker (1999)
  \item \code{B = 0.67}, fit parameter for drop velocity from diameters, 
  see Uijlenhoet and Stricker (1999)
  \item \code{epsilon = 1}, impact elasticity, \code{1} for liquid water
  \item \code{v_max = 9.5}, maximum fall velocity for rain drops (m/s), 
  all higher velocities will be set to this value
  \item \code{r_w = 1000}, specific density of the rain drops (kg/m^3)
  \item \code{f_0 = 1}, reference frequency (Hz), used to adjust frequency
  dependent quality factor and wave velocity values
  \item \code{alpha = -0.85}, constant for Rayleigh wave particle motion 
  attenuation with depth from surface, see Stein and Wysession (2003) 
  (Eq. 2.7-9) for details.
  \item \code{beta = -0.39}, constant for Rayleigh wave particle motion 
  attenuation with depth from surface, see Stein and Wysession (2003) 
  (Eq. 2.7-9) for details.
  
}
}
\examples{

## model rainfall spectrum for 40 mm/h intensity, sensor at 0.5 m depth
p <- model_rainfall(I = 40, 
                    q_0 = 20, 
                    v_0 = 595, 
                    e_0 = 0.2,
                    p_0 = 1, 
                    depth = 0.5)

## plot spectrum
plot(p)

## model rainfall spectrum with manually defined rain properties
p <- model_rainfall(I = 40, 
                    D_size = c(0.1, 0.2, 0.5, 1, 2, 5, 10),
                    D_prob = c(1e-1, 1e-2, 1e-3, 1e-5, 1e-7, 1e-9, 5e-11),
                    D_velo = c(0.8, 5.5, 8.5, 9.5, 9.5, 9.5, 9.5),
                    q_0 = 20, 
                    v_0 = 595, 
                    e_0 = 0.2,
                    p_0 = 1, 
                    depth = 0.5)
                    
## plot spectrum
plot(p)

}
\references{
Stein S and Wysession M (2003). An Introduction to Seismology, Earthquakes, 
and Earth Structure. Blackwell. ISBN-13: 978-0-86542-078-6 

Uijlenhoet R and Stricker JNM (1999). A consistent rainfall parameterization 
based on the exponential raindrop size distribution. Journal of Hydrology 
218, 101-127. DOI: 10.1016/S0022-1694(99)00032-3.

Tsai VC and Atiganyanun S (2014). Greens functions for surface waves in a 
generic velocity StructureShort note. Bulletin of the Seismological 
Society of America, 104(5), 2573-2578. DOI: 10.1785/0120140121.

Tsai VC, Minchew B, Lamb MP, and Ampuero J-P (2012). A physical model for
seismic noise generation from sediment transport in rivers, Geophysical
Research Letters 39, L02404. DOI: 10.1029/2011GL050255.
}
\author{
Ryan Porter, Michael Dietze
}
\keyword{eseis}
