% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlogram.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Plot a correlogram from noise cross correlation analysis}
\usage{
plot_correlogram(
  data,
  ratio = 0.3,
  agg = c(1, 1),
  legend = TRUE,
  keep_par = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{List} object, spectrogram to be plotted. Must be output
of \code{ncc_correlate()} or of equivalent structure.}

\item{ratio}{\code{Numeric} value, ratio of the space that is taken by 
the averaged correlation function in favour of the correlogram. Default
is \code{0.3} (30 percent width allocated to averaged correlation 
function, and 70 % allocated to the correlogram). If set to \cdoe{0}, 
only the correlogram is plotted. If set to \code{1}, only the correlation
function is plotted.}

\item{agg}{\code{Integer} vector of length two, factors of image 
aggregation, i.e. in time and lag dimension. Useful to decrease 
image size. Default is \code{c(1, 1)} (no aggregation).}

\item{legend}{\code{Logical} value, option to add colour bar legend. Legend
label can be changed by \code{zlab}.}

\item{keep_par}{\code{Logical} value, option to omit resetting plot 
parameters after function execution. Useful for adding further data to the 
plot. Default is \code{FALSE} (parameters are reset to original values).}

\item{\dots}{Additional arguments passed to the plot function.}
}
\value{
Graphic output of a correlogram.
}
\description{
The function uses the output of \code{ncc_correlate()} to show a 
compound plot with a time averaged correlation function and its 
inter-quartile range polygon next to an image plot of a noise cross 
correlation functions resolved by time. The geometric ratio of the two
plots can be controlled, and each plot can also be omitted entirely.
}
\examples{

\dontrun{
  
  ## calculate correlogram
  cc <- ncc_correlate(start = "2017-04-09 00:30:00", 
                       stop = "2017-04-09 01:30:00", 
                       ID = c("RUEG1", "RUEG2"), 
                       component = c("Z", "Z"), 
                       dir = paste0(system.file("extdata", 
                                    package = "eseis"), "/"), 
                       window = 600, 
                       overlap = 0, 
                       lag = 20, 
                       f = c(0.05, 0.1), 
                       sd = 1)
                       
   ## explicit plot function call with adjusted resolution
   plot_correlogram(data = cc, agg = c(2, 5))
   
   ## define plot colour scale
   cls <- colorRampPalette(colors = c("brown", "white", "green"))
   
   ## simple function call with user-defined colour scale
   plot(cc, col = cls(100))
   
   ## show plot without time averaged correlation data
   plot(cc, ratio = 0)
   
   ## show plot without time resolved correlation data
   plot(cc, ratio = 1)
}

}
\seealso{
\code{\link{ncc_correlate}}
}
\author{
Michael Dietze
}
\keyword{eseis}
