% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dvv.R
\name{plot_dvv}
\alias{plot_dvv}
\title{Plot dv/v results from noise cross correlation analysis}
\usage{
plot_dvv(data, ratio = 0.5, n, points = TRUE, keep_par = FALSE, ...)
}
\arguments{
\item{data}{\code{List} object, spectrogram to be plotted. Must be output
of \code{ncc_stretch()} or of equivalent structure.}

\item{ratio}{\code{Numeric} value, ratio of the space that is taken by 
the time series in favour of the goodness-of-fit matrix plot. Default is 
\code{0.5} (50 percent width allocated to dv/v time series and 50 percent
allocated to the goodness-of-fit matrix). If set to \code{0}, only the 
goodness-of-fit matrix is plotted. If set to \code{1}, only the dv/v time 
series is plotted}

\item{n}{\code{Numeric} value, optional number of samples in 
running window used for smoothing the spectrogram. Only applied if a 
number is provided. Smoothing is performed as running mean.}

\item{points}{\code{Logical} value, option to add white points to 
goodness-of-fit matrix, where maximum fit quality is located. The 
transparency of the points expresses the strength of the goodness-of-fit. 
Default is \code{TRUE}.}

\item{keep_par}{\code{Logical} value, option to omit resetting plot 
parameters after function execution. Useful for adding further data to the 
plot. Default is \code{FALSE} (parameters are reset to original values).}

\item{\dots}{Additional arguments passed to the plot function.}
}
\value{
Graphic output of a relative seismic wave velocity change estimate
}
\description{
The function uses the output of \code{ncc_stretch()} to show a 
compound plot with a relative seismic wave velocity change (dv/v) time 
series together with the goodness-of-fit matrix for all tested dvv values.
}
\details{
By default the goodness-of-fit matrix will be scaled to the full range of 
tested dv/v values, whereas the dv/v time series will be scaled to the 
actual range of dv/v values. By setting \code{ylim}, both plots will be 
scaled to that range.
}
\examples{

\dontrun{
  
  ## calculate correlogram
  cc <- ncc_correlate(start = "2017-04-09 00:30:00", 
                      stop = "2017-04-09 01:30:00", 
                      ID = c("RUEG1", "RUEG2"), 
                      dt = 1/10,
                      component = c("Z", "Z"), 
                      dir = paste0(system.file("extdata", 
                                   package = "eseis"), "/"), 
                      window = 600, 
                      overlap = 0, 
                      lag = 20, 
                      f = c(0.05, 0.1), 
                      sd = 1)
   
   ## estimate dv/v
   dv <- ncc_stretch(data = cc, 
                     lag = c(1, 10), 
                     range = 0.4)
   
   ## plot result
   plot_dvv(data = dv)
   
   ## plot only goodness-of-fit matrix
   plot_dvv(data = dv, ratio = 0)
   
   ## plot only dv/v time series
   plot_dvv(data = dv, ratio = 1)
   
   ## plot both again but with different plot ratio
   plot_dvv(data = dv, ratio = 0.6)
   
   ## plot matrix with different colour scale
   plot_dvv(data = dv, col = heat.colors(200))
   
   ## plot with smoothened dv/v curve
   plot_dvv(data = dv, n = 3)
   
}

}
\seealso{
\code{\link{ncc_stretch}}, \code{\link{ncc_correlate}}
}
\author{
Michael Dietze
}
\keyword{eseis}
