% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mseed.R
\name{read_mseed}
\alias{read_mseed}
\title{Read mseed file.}
\usage{
read_mseed(
  file,
  signal = TRUE,
  time = TRUE,
  meta = TRUE,
  header = TRUE,
  eseis = TRUE,
  type = "waveform"
)
}
\arguments{
\item{file}{\code{Character} value, input file name, with extension.}

\item{signal}{\code{Logical} value, option to import the signal vector,
default is \code{TRUE}.}

\item{time}{\code{Logical} value, option to create the time vector. The
timezone is automatically set to \code{"UTC"}, default is \code{TRUE}.}

\item{meta}{\code{Logical} value, option to append the meta data part,
default is \code{TRUE}.}

\item{header}{\code{Logical} value, option to append the header part,
default is \code{TRUE}.}

\item{eseis}{\code{Logical} value, option to read data to an \code{eseis}
object (recommended, see documentation of
\code{aux_initiateeseis}), default is \code{TRUE}}

\item{type}{\code{Character} value, type keyword of the data. One out of
\code{"waveform"}, \code{"envelope"}, \code{"fft"}, \code{"spectrum"},
\code{"spectrogram"}, \code{"other"}, \code{hilbert}, \code{hvratio}.
Default is \code{"waveform"}.}
}
\value{
\code{List} object, optionally of class \code{eseis}
}
\description{
The function reads mseed files.
}
\details{
When a file contains more than one trace per channel, the function assumes
the sampling rate is identical and only extracts the sampling rate of the 
first trace as output sampling rate. In that case, the output is not one 
\code{eseis} object (or list object in case option \code{eseis = FALSE}), 
but a list of \code{eseis} objects named by their component code.  

The mseed data format is read using the C library libmseed v.3.2.3
(https://github.com/earthscope/libmseed). The C wrapper for R has been 
build with the help of ChatGPT.
}
\examples{

\dontrun{

## read mseed file with default options
x <- read_mseed(file = "input.miniseed")

## read mseed file, only signal trace, not as eseis object
x <- read_mseed(file = "input.miniseed",
                time = FALSE,
                meta = FALSE,
                header = FALSE,
                eseis = FALSE)
                
}

}
\author{
Michael Dietze
}
