% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{data}
\name{rockfall}
\alias{rockfall}
\title{Seismic trace of a rockfall event.}
\format{The format is: num [1:98400] 65158 65176 65206 65194 65155 ...}
\usage{
rockfall
}
\description{
The dataset comprises the seismic signal (vertical component) of 
a rockfall event, preceeded by an earthquake. The data have been
recorded at 200 Hz sampling frequency with an Omnirecs Cube ext 3
data logger.
}
\examples{

## load example data set
data(rockfall)

## plot signal vector
plot(x = t, y = rockfall, type = "l")

}
\keyword{datasets}
