\name{sac}
\alias{sac}
\docType{data}
\title{example data}
\description{Seismic data set in sac-format.}
\usage{data(sac)}
\format{
  The format is:
List of 1
 $ :List of 10
  ..$ fn    : chr "IG008.13.001.11.00.00.BHZ.SAC"
  ..$ sta   : chr "IG008"
  ..$ comp  : chr "BHZ"
  ..$ dt    : num 0.005
  ..$ DATTIM:List of 12
  .. ..$ yr  : num 2013
  .. ..$ jd  : num 1
  .. ..$ mo  : num 1
}
\details{The dataset was recorded by a Trillium Compact broadband seismometer.}
\examples{
## load example data set
data(sac)
}
\keyword{datasets}
