\name{setup.1}
\alias{setup.1}
\docType{data}
\title{example data}
\description{Example setup information for a TC120 and an Omnirecs Cube 3.}
\usage{data(setup.1)}
\format{
  The format is:
List of 10
 $ ID         : chr "Station 01"
 $ type.sensor: chr "Trillium Compact 120 s"
 $ type.logger: chr "Data Cube 3 ext"
 $ coordinates: num [1:3] 5395031 5651928 600
 $ poles      : cplx [1:11] -0.0369+0.037i -0.0369-0.037i -343+0i ...
 $ zeros      : cplx [1:6] 0+0i 0+0i -392+0i ...
 $ k          : num 4.34e+17
 $ s          : num 754
 $ gain       : num 1
 $ AD         : num 2.44e-07
}
\details{No details, yet.}
\examples{
## load example data set
data(setup.1)
}
\keyword{datasets}
