% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_clip.R
\name{signal_clip}
\alias{signal_clip}
\title{Clip signal based on time vector.}
\usage{
signal_clip(data, time, limits)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors, input signal vector.}

\item{time}{\code{POSIXct} vector, corresponding time vector.}

\item{limits}{\code{POSIXct} vector of length two, time limits for 
clipping.}
}
\value{
\code{Numeric} data set clipped to provided time interval.
}
\description{
The function clips signals based on the correpsonding time vector.
}
\examples{

## TO BE DONE
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
