% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_deconvolve.R
\name{signal_deconvolve}
\alias{signal_deconvolve}
\title{Deconvolve a signal vector.}
\usage{
signal_deconvolve(data, dt, sensor = "TC120s", logger = "Cube3extBOB",
  gain = 1, p = 10^-6, waterlevel = 10^-6, na.replace = FALSE)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}

\item{dt}{\code{Numeric} scalar, sampling rate.}

\item{sensor}{\code{Character} scalar or \code{list} object, 
seismic sensor name. Must be present in the sensor library 
(\code{list_sensor}) or parameters must be added manually (see 
examples). Default is \code{"TC120s"}.}

\item{logger}{\code{Character} scalar, seismic logger name. Must be 
present in the logger library (\code{list_logger}) or parameters must be
added manually. Default is \code{"Cube3extBOB"}.}

\item{gain}{\code{Numeric} value, signal gain level of the logger. Default 
is \code{1}.}

\item{p}{\code{Numeric} scalar, proportion of signal to be tapered. Default
is\code{10^-6}.}

\item{waterlevel}{\code{Numeric} scalar, waterlevel value for frequency
division, default is \code{10^-6}.}

\item{na.replace}{\code{Logical} value, option to replace NA values in the 
data set by zeros. Default is \code{FALSE}. Attention, the zeros will 
create artifacts in the deconvolved data set. However, NA values will 
result in no deconvolution at all.}
}
\value{
\code{Numeric} vector or list of vectors, deconvolved signal.
}
\description{
The function removes the instrument response from a signal vector.
}
\details{
The function requires a set of input parameters, apart from the signal 
vector. These parameters are contained in and read from the function 
\code{list_sensor()} and \code{list_logger()}. Poles and zeros are used 
to build the transfer function. The value s is the generator constant in 
Vs/m. The value k is the normalisation factor. AD is the analogue-digital 
conversion factor. If the signal was recorded with a gain value other than 
1, the resulting signal needs to be corrected for this, as well.
}
\examples{

## load example data set
data(rockfall)

## deconvolve signal
rockfall_decon <- signal_deconvolve(data = rockfall_z,
                                    dt = 1/200, 
                                    sensor = "TC120s", 
                                    logger = "Cube3ext")

## plot time series
plot(x = rockfall_t, 
     y = rockfall_decon, 
     main = "Rockfall, deconvolved signal", 
     ylab = "m/s", 
     type = "l")
 
## add new logger manually
logger_new <- list_logger()[[1]]

## add logger data
logger_new$ID <- "logger_new"
logger_new$name <- "logger_new"
logger_new$AD <- 2.4414e-07

## deconvolve signal with new logger
rockfall_decon <- signal_deconvolve(data = rockfall_z,
                                    dt = 1/200, 
                                    sensor = "TC120s", 
                                    logger = logger_new)
                                    
## Change the setup of a logger, here: Centaur AD is changed due to 
## other than default Vpp value, according to AD = V / (2^24).

## extract default Centaur logger
Centaur_10V <- list_logger()[[2]]

## replace AD value
Centaur_10V$AD <- 20/(2^24)
                                    
}
\author{
Michael Dietze
}
\keyword{eseis}
