% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_demean.R
\name{signal_demean}
\alias{signal_demean}
\title{Remove mean of signal vector.}
\usage{
signal_demean(data)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}
}
\value{
\code{Numeric} vector or list of vectors, data set with mean 
subtracted.
}
\description{
The function removes the mean from a signal vector.
}
\examples{

## load example data set
data(rockfall)

## remove mean from data set
rockfall_demean <- signal_demean(data = rockfall_z)

## compare data ranges
range(rockfall_z)
range(rockfall_demean)

## show mean of initial signal
mean(rockfall_z)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
