% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_filter.R
\name{signal_filter}
\alias{signal_filter}
\title{Filter a seismic signal.}
\usage{
signal_filter(data, dt, f, type, shape = "butter", order = 2, p = 0)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors, input signal vector.}

\item{dt}{\code{Numeric} scalar, sampling period. If omitted, \code{dt} 
is set to 1/200.}

\item{f}{\code{Numeric} scalar or vector of length two, lower and/or 
upper cutoff frequencies (Hz).}

\item{type}{\code{Character} scalar, type of filter, one out of 
\code{"LP"} (low pass), \code{"HP"} (high pass), \code{"BP"} (band 
pass) and \code{"BR"} (band rejection). If omitted, the type is interpreted 
from \code{f}. If \code{f} is of length two, \code{type} is set to 
\code{"BP"}. If \code{f} is of length one, \code{type} is set to 
\code{"HP"}.}

\item{shape}{\code{Character} scalar, one out of \code{"butter"} 
(Butterworth), default is \code{"butter"}.}

\item{order}{\code{Numeric} scalar, order of the filter, default 
is \code{2}.}

\item{p}{\code{Numeric} scalar, fraction of the signal to be tapered.}
}
\value{
\code{Numeric} vector or list of vectors, filtered signal vector.
}
\description{
The function filters the input signal vector.
}
\examples{

## load example data set
data(rockfall)

## filter data set by bandpass filter between 1 and 90 Hz
rockfall_bp <- signal_filter(data = rockfall_z, 
                             f = c(1, 90), 
                             dt = 1/200)
                             
## taper signal to account for edge effects
rockfall_bp <- signal_taper(data = rockfall_bp, n = 2000)

## plot filtered signal
plot(x = rockfall_t[68000:78000], 
     y = rockfall_bp[68000:78000], 
     type = "l")
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
