% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_hilbert.R
\name{signal_hilbert}
\alias{signal_hilbert}
\title{Calculate Hilbert transform.}
\usage{
signal_hilbert(data)
}
\arguments{
\item{data}{\code{Numeric} vector, input signal vector.}
}
\value{
\code{Numeric} vector, Hilbert transform.
}
\description{
The function calculates the Hilbert transform of the input signal vector.
}
\examples{

## NOT DONE YET
                     
}
\author{
Michael Dietze
}
\keyword{eseis}

