% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_pad.R
\name{signal_pad}
\alias{signal_pad}
\title{Pad signal with zeros.}
\usage{
signal_pad(data)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}
}
\value{
\code{Numeric} vector or list of vectors, signal vector with 
added zeros.
}
\description{
The function adds zeros to the input vector to reach a length,
corresponding to the next higher power of two.
}
\examples{

## load example data set
data(rockfall)

## pad with zeros
rockfall_pad <- signal_pad(data = rockfall_eseis)

## compare lengths
rockfall_eseis$meta$n
rockfall_pad$meta$n
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
