% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sign.R
\name{signal_sign}
\alias{signal_sign}
\title{Convert amplitude signal to one bit signed signal}
\usage{
signal_sign(data)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}
}
\value{
\code{Numeric} vector or list of vectors, sign-cut signal.
}
\description{
This function assigns 1 for positive values and -1 for negative input
values of a signal.
}
\examples{

## load example data
data(rockfall)

## sign-cut signal
rockfall_sign <- signal_sign(data = rockfall_eseis)

}
\author{
Michael Dietze
}
\keyword{eseis}
