% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_snr.R
\name{signal_snr}
\alias{signal_snr}
\title{Calculate signal-to-noise-ratio.}
\usage{
signal_snr(data, detrend = FALSE)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors , input signal 
vector.}

\item{detrend}{\code{Logical} scalar, optionally detrend data set before
calcualting snr.}
}
\value{
\code{Numeric} scalar, signal-to-noise ratio.
}
\description{
The function calculates the signal-to-noise ratio of an input signal 
vector as the ratio between mean and max.
}
\examples{

## load example data set
data(rockfall)

## calculate snr with detrend option off and on
signal_snr(data = rockfall)
signal_snr(data = rockfall, detrend = TRUE)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
