% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_spectrogram.R
\name{signal_spectrogram}
\alias{signal_spectrogram}
\title{Calculate spectrograms (power spectral density estimates) from time series.}
\usage{
signal_spectrogram(data, time, dt, Welch = FALSE, window, overlap = 0.5,
  window_sub, overlap_sub = 0.5, multitaper = FALSE, nw = 4, k = 7,
  n_cores = 1, plot = FALSE)
}
\arguments{
\item{data}{\code{Numeric} vector, seismic signal to be processed.}

\item{time}{\code{POSIX.ct} vector with time values. If omitted, an 
artificial time vector will be created, based on \code{dt}.}

\item{dt}{\code{Numeric} scalar, sampling period. If omitted, either 
estimated from \code{time} or set to 0.01 s (i.e., f = 100 Hz).}

\item{Welch}{\code{Logical} scalar, option to use the Welch method for
calcualtions.}

\item{window}{\code{Numeric} scalar, time window length used to 
calculate individual spectrae. Set to 1 percent of the time series length 
by default.}

\item{overlap}{\code{Numeric} scalar, fraction of window overlap.}

\item{window_sub}{\code{Numeric} scalar, length of the sub-window used 
to calculate spectrae. Only relevant if \code{Welch = TRUE}. If omitted, 
the sub-window length is set to 10 percent of the main window length.}

\item{overlap_sub}{\code{Numeric} scalar, fraction of sub-window overlap.}

\item{multitaper}{\code{Logical} scalar, option to use multitaper for the
calculations. Can increase computation time significantly.}

\item{nw}{\code{Numeric} scalar, multitaper time-bandwidth parameter,
default is 4.0.}

\item{k}{\code{Numeric} scalar, multitaper number of tapers, default is 7.}

\item{n_cores}{\code{Numeric} scalar, number of CPU cores to use. Disabled 
by setting to 1. Default is 1.}

\item{plot}{\code{logical} scalar, toggle plot output. Default is
\code{FALSE}.}
}
\value{
\code{List} with spectrogram matrix, time and frequency vectors.
}
\description{
This function creates spectrograms from seismic signals. It supports the
standard spectrogram apporach, multitaper, and the Welch method.
}
\examples{

## TO BE ADDED LATER
# not run
                     
}
\author{
Michael Dietze
}
\seealso{
\code{\link{spectrum}}, \code{\link{spec.pgram}}, 
\code{\link{spec.mtm}}
}
\keyword{eseis}

