% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_stalta.R
\name{signal_stalta}
\alias{signal_stalta}
\title{Calculate stal-lta-ratio.}
\usage{
signal_stalta(data, sta, lta)
}
\arguments{
\item{data}{\code{Numeric} vector or list , input signal envelope vector(s)}

\item{sta}{\code{Numeric} scalar, number of samples for short-term window.}

\item{lta}{\code{Numeric} scalar, number of samples for long-term window.}
}
\value{
\code{Numeric} vector, sta-lta-ratio.
}
\description{
The function calculates the ratio of the short-term-average and 
long-term-average of the input signal.
}
\examples{

## load example data set
data(rockfall)

## TO BE DONE, LATER
                     
}
\author{
Michael Dietze
}
\keyword{eseis}

