% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_sum.R
\name{signal_sum}
\alias{signal_sum}
\title{Calculate signal vector sum.}
\usage{
signal_sum(...)
}
\arguments{
\item{\dots}{\code{Numeric} vectors or \code{eseis} objects, input 
signal, that must be of the same length.}
}
\value{
\code{Numeric} vector, signal vector sum.
}
\description{
The function calculates the vector sum of the input signals.
}
\examples{

## create random vectors
x <- runif(n = 1000, min = -1, max = 1)
y <- runif(n = 1000, min = -1, max = 1)
z <- runif(n = 1000, min = -1, max = 1)

## calculate vector sums
xyz <- signal_sum(x, y, z)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
