\name{snr}
\alias{snr}
\title{Function to calculate signal-to-noise-ratios.}
\description{The function returns the signal-to-noise-ratios (SNR) of input signals. 
The SNR is defined as max(data) / mean(data).}
\usage{snr(data)}
\arguments{
  \item{data}{Numeric \code{vector} or \code{matrix} with respective seismic traces.}
}

\value{A vector of time units.}

\author{Michael Dietze}




\examples{
## no example yet

}

\keyword{environmental}
\keyword{seismology}
