% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_convert.R
\name{spatial_convert}
\alias{spatial_convert}
\title{Convert coordinates between reference systems}
\usage{
spatial_convert(data, from, to)
}
\arguments{
\item{data}{\code{Numeric} matrix or data frame, x-, y-coordinates to 
be converted.}

\item{from}{\code{Character} scalar, proj4 string of the current 
reference system.}

\item{to}{\code{Character} scalar, proj4 string of the target 
reference system.}
}
\value{
\code{Numeric} data frame with converted coordinates.
}
\description{
Coordinates are converted between reference systems.
}
\examples{

xy <- cbind(13, 55)

proj_in <- "+proj=longlat + ellps=WGS84"

proj_out <- "+proj=utm +zone=32 +datum=WGS84"

spatial_convert(data = xy, 
                from = proj_in, 
                to = proj_out)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}

