% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_migrate.R
\name{spatial_migrate}
\alias{spatial_migrate}
\title{Migrate signals of a seismic event through a grid of locations.}
\usage{
spatial_migrate(data, d_stations, d_map, snr, v, dt, normalise = TRUE)
}
\arguments{
\item{data}{\code{Numeric} matrix, seismic signals to cross-correlate.}

\item{d_stations}{\code{Numeric} matrix, inter-station distances. Output 
of \code{distance_stations}.}

\item{d_map}{\code{List} object, distance maps for each station (i.e., 
\code{SpatialGridDataFrame} objects). Output of \code{distance_map}.}

\item{snr}{\code{Numeric} vector, optional signal-to-noise-ratios for 
each signal trace, used for normalisation. If omitted it is calculated
from input signals.}

\item{v}{\code{Numeric} value, mean velocity of seismic waves (m/s).}

\item{dt}{\code{Numeric} value, sampling period.}

\item{normalise}{\code{Logical} value, option to normalise stations 
correlations by signal-to-noise-ratios.}
}
\value{
A SpatialGridDataFrame-object with Gaussian probability density
function values for each grid cell.
}
\description{
The function performs signal migration in space in order to determine 
the location of a seismic signal.
}
\examples{

## No examples ready, yet.

}
\author{
Michael Dietze
}
