% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{data}
\name{t}
\alias{t}
\title{example data}
\format{The format is: POSIXct[1:98400], format: "2015-04-06 13:16:54" ...}
\description{
Time vector of a rockfall event.
}
\details{
The dataset comprises the time vector corresponding the to seismic signal
of a rockfall event.
}
\examples{

## load example data set
data(examples)

}
\keyword{datasets}

