% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_aggregate.R
\name{time_aggregate}
\alias{time_aggregate}
\title{Aggregate a time series}
\usage{
time_aggregate(data, n = 2)
}
\arguments{
\item{data}{\code{Numeric} vector, time to be processed.}

\item{n}{\code{Numeric} scalar, number of samples to be aggregated to one
new data value. Must be an integer value greater than 1. Default is 
\code{2}.}
}
\value{
\code{Numeric} vector, aggregated data.
}
\description{
The time series \code{x} is aggregated by an integer factor \code{n}.
}
\examples{

## load example data set
data(rockfall, envir = environment())

## aggregate time series
t_agg <- time_aggregate(data = t, 
                          n = 2)

## compare results
range(t)
diff(t)

range(t_agg)
diff(t_agg)

}
\author{
Michael Dietze
}
\keyword{eseis}

