% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sac.R
\name{write_sac}
\alias{write_sac}
\title{Write seismic traces as sac-file to disk.}
\usage{
write_sac(data, file, time, component, unit, station, location, network, dt,
  parameters, biglong = FALSE)
}
\arguments{
\item{data}{\code{Numeric} vector, seismic trace to be written.}

\item{file}{\code{Character} scalar, filename with extension.}

\item{time}{\code{POSIXct} vector, time vector corresponding to the 
seismic trace. Alternatively, the start time stamp can be provided as
\code{POSIXct} value and a value for \code{dt} must be given.}

\item{component}{\code{Character} scalar, component ID, optional.}

\item{unit}{\code{Character} scalar, unit of the signal, optional. One out
of \code{"unknown"}, \code{"displacement"}, \code{"velocity"}, 
\code{"volts"}, \code{"acceleration"}. Default is \code{"unknown"}.}

\item{station}{\code{Character} scalar, station ID, optinal.}

\item{location}{\code{Character} vector of length four, station location 
data (latitude, longitude, elevation, depth), optional.}

\item{network}{\code{Character} scalar, network ID, optional.}

\item{dt}{\code{Numeric} scalar, sampling period. Only needed if no time
vector is provided.}

\item{parameters}{\code{Data frame} sac parameter list, as obtained from
\code{list_sacparameters}. Allows user-specific modifications. If this 
data frame is provided, it overrides all other arguments.}

\item{biglong}{\code{Logical} scalar, biglong option, default is 
\code{FALSE}}
}
\value{
A binary file written to disk.
}
\description{
This function converts seismic traces to sac-files and writes them to disk.
}
\details{
For description of the sac file format see
https://ds.iris.edu/files/sac-manual/manual/file_format.html. Currently the 
following parameters are not supported when writing the sac file: 
LAT, LON, ELEVATION, NETWORK.
}
\examples{

## No examples ready, yet.

}
\author{
Michael Dietze
}

