% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_clipspectrogram.R
\name{aux_clipspectrogram}
\alias{aux_clipspectrogram}
\title{Clip, interpolate (and optionally smoothen) a spectrogram}
\usage{
aux_clipspectrogram(data, f, res = 100, n, dt)
}
\arguments{
\item{data}{\code{List} object, spectrogram to be processed. Must be the 
output of \code{signal_spectrogram} or of an equivalent structure.}

\item{f}{\code{Numeric} vector of length two, frequency range to which 
the PSD will be clipped. Alternatively, an empirical frequency vector can 
be submitted, too (e.g., when non-uniformly spaced frequencies are used).}

\item{res}{\code{Numeric} value, number of frequencie values used as output 
of the frequency interpolation. If set to \code{1}, 
the input frequency range (\code{f}) will be averaged and a 
\code{data frame} with the elements \code{time} and \code{power} will 
be returned.}

\item{n}{\code{Numeric} value, number of values to use in a running average 
filter to smoothen the interpolated output. If omitted, the spectrogram 
will not be changed apart from the interpolation option.}

\item{dt}{\code{Numeric} value, sampling interval. Only required if 
\code{data} is no eseis object.}
}
\value{
A clipped and interpolated spectrogram, or a data frame with 
a time series of an averaged frequency band.
}
\description{
The function interpolates the power values in the frequency range defined 
by the user, applying a spline interpolator. This function can be used to 
extract a given frequency range or frequency band from a PSD. The former is 
useful for convenient preparation of the \code{fmi_} function family to 
achieve fluvial model inversion workflows. The latter is useful to create 
seismic power time series for a given frequency band.
}
\examples{

## load example data
data(rockfall)

## calculate spectrogram
psd <- signal_spectrogram(data = rockfall_eseis)

## Example 1: interpolate PSD to 5 to 15 Hz with 100 values resolution
psd_int <- aux_clipspectrogram(data = psd, f = c(5, 15), res = 100)

## plot output
plot(psd_int)

## Example 2: interpolate PSD using user defined frequency vector
psd_int <- aux_clipspectrogram(data = psd, f = sort(runif(n = 20, 5, 20)))

## plot output
plot(psd_int)

## Example 3: interpolate non-eseis matrix object
psd_int <- aux_clipspectrogram(data = psd$PSD$S, 
                               f = c(0, 10), 
                               res = 10, 
                               dt = 1/50)
                               
## Example 4: extract a power time series for a frequency band
p_ts <- aux_clipspectrogram(data = psd, f = c(5, 10), res = 1)
plot(p_ts, type = "l")

## Example 5: clip and smoothen the output
p_sm <- aux_clipspectrogram(data = psd, f = c(5, 20), res = 50, n = 10)
plot(p_sm)

}
\author{
Michael Dietze
}
\keyword{eseis}
