\name{eseis-package}
\alias{eseis-package}
\alias{eseis}
\docType{package}
\title{Environmental seismology toolbox}
\description{This package provides a set of functions for handling seismic data to monitor Earth surface processes.}
\details{
\tabular{ll}{Package: \tab eseis\cr
Type: \tab Package\cr
Title: \tab Environmental seismology toolbox\cr
Version: \tab 0.2.3\cr
Date: \tab 2015-07-15\cr
Author: \tab Michael Dietze\cr
Maintainer: \tab Michael Dietze <mdietze@gfz-potsdam.de>\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 3.1.1), caTools, fields, RSEIS, signal, zoo, rPython\cr}
}
\author{Michael Dietze}

\keyword{ package }