% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{package}
\name{eseis-package}
\alias{eseis}
\alias{eseis-package}
\title{eseis}
\description{
Environmental seismology toolbox
}
\details{
\tabular{ll}{ Package: \tab MDtools\cr Type: \tab Package\cr Version: \tab
0.3.0\cr Date: \tab 2016-06-17\cr License: \tab GPL-3\cr }
}
\author{
Michael Dietze
}
\keyword{package}

