% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{package}
\name{eseis-package}
\alias{eseis-package}
\alias{eseis}
\title{eseis: Environmental Seismology Toolbox}
\description{
Environmental seismoloy studies the seismic
signals, emitted by Earth surface processes. This package eseis provides
all relevant functions to read/write seismic data files, prepare, analyse
and visualise seismic data, and generate reports of the processing history.
}
\author{
\strong{Maintainer}: Michael Dietze \email{michael.dietze@uni-goettingen.de} [translator]

Authors:
\itemize{
  \item Clement Hibert [contributor]
}

Other contributors:
\itemize{
  \item Christoph Burow [contributor]
  \item Sophie Lagarde [contributor, translator]
}

}
\keyword{package}
