% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{package}
\name{eseis}
\alias{eseis}
\alias{eseis-package}
\title{eseis: Environmental Seismology Toolbox}
\description{
Environmental seismoloy is a scientific field that studies the seismic 
signals, emitted by Earth surface processes. This package eseis provides 
all relevant functions to read/write seismic data files, prepare, analyse
and visualise seismic data, and generate reports of the processing history.
}
\details{
\tabular{ll}{
**Package:** \tab eseis \cr
**Type:** \tab Package \cr
**Version:** \tab 0.4.0 \cr
**Date:** \tab 2018-04-25 \cr
**License:** \tab GPL-3 \cr
}
}
\author{
Michael Dietze
}
\keyword{package}
