% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_models.R
\name{gui_models}
\alias{gui_models}
\title{Start GUI with seismic models}
\usage{
gui_models(...)
}
\arguments{
\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\description{
This function starts a browser-based graphic user interface to explore the 
parameter space of seismic models that predict the spectra of turbulent 
water flow and bedload flux.
}
\examples{

\dontrun{
# Start the GUI
gui_models()
}

}
\seealso{
\code{\link{runApp}}
}
\author{
Michael Dietze
}
