% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_logger.R
\name{list_logger}
\alias{list_logger}
\title{List library with data logger information.}
\usage{
list_logger()
}
\value{
\code{List} object, supported loggers with their parameters.
}
\description{
The function returns the list of supported data loggers to extract signal 
deconvolution parameters.
}
\details{
The value AD is the analogue-digital conversion factor.
}
\examples{

## show documented loggers
list_logger()

## show names of loggers in list
names(list_logger())
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
