% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sacparameters.R
\name{list_sacparameters}
\alias{list_sacparameters}
\title{List all header parameters of a sac file.}
\usage{
list_sacparameters()
}
\value{
\code{List} object, parameters supported by a sac file.
}
\description{
The function returns a data frame with all header values of a sac file. It 
may be used for advanced modifications by the user.
}
\examples{

## show sac parameters
list_sacparameters()
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
