% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sensor.R
\name{list_sensor}
\alias{list_sensor}
\title{List sensor library.}
\usage{
list_sensor()
}
\value{
\code{List} object, supported sensors with their parameters.
}
\description{
The function returns the list of supported sensors to extract signal 
deconvolution parameters.
}
\details{
Poles and zeros must be given in rad/s. Characteristics of further 
sensors can be added manually. See examples of \code{signal_deconvolve}
for further information. The value s is the generator constant 
(sensitivity) given in Vs/m. The value k is the normalisation factor of 
the sensor.
}
\examples{

## show sensors
list_sensor()
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
