% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrogram.R
\name{plot_spectrogram}
\alias{plot_spectrogram}
\title{Calculate spectrograms (power spectral density estimates) from time series.}
\usage{
plot_spectrogram(data, col = "gradient_1")
}
\arguments{
\item{data}{\code{List} object, spectrogram to be plotted. Must be output
of \code{signal_spectrogram()} or of equivalent structure.}

\item{col}{\code{Character} scalar, colour palette to use. Default is 
\code{tim.colors}.}
}
\value{
Graphic output of a spectrogram.
}
\description{
This function creates spectrograms from seismic signals. It supports the
standard spectrogram apporach, multitaper, and the Welch method.
}
\examples{

## load example data set
data(rockfall)

## calculate spectrogram
PSD <- signal_spectrogram(data = rockfall, time = t, dt = 1/200)

## plot spectrogram
plot_spectrogram(data = PSD)
                     
}
\author{
Michael Dietze
}
\seealso{
\code{\link{signal_spectrogram}}
}
\keyword{eseis}

