% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{data}
\name{rockfall}
\alias{rockfall}
\title{example data}
\format{The format is: num [1:98400] 65158 65176 65206 65194 65155 ...}
\description{
Seismic trace of a rockfall event.
}
\details{
The dataset comprises the seismic signal (vertical component) of a rockfall
event, preceeded by an earthquake.
}
\examples{

## load example data set
data(rockfall)

## plot signal vector
plot(x = t, y = rockfall, type = "l")

}
\keyword{datasets}

