% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_deconvolve.R
\name{signal_deconvolve}
\alias{signal_deconvolve}
\title{Deconvolve a signal vector.}
\usage{
signal_deconvolve(data, dt, output = "velocity", sensor = "TC120s",
  logger = "Cube3ext", p = 10^-6, waterlevel = 10^-6)
}
\arguments{
\item{data}{\code{Numeric} vector, input signal vector}

\item{dt}{\code{Numeric} scalar, sampling rate.}

\item{output}{\code{Character} scalar, output data type. One out 
of \code{"velocity"}, \code{"displacement"} or \code{"acceleration"}, 
default is \code{"velocity"}. The other keywords are not yet supported.}

\item{sensor}{\code{Character} scalar, seismic sensor name. Must be 
present in the sensor library (\code{list_sensor}) or parameters must be 
added manually. Default is \code{"TC120s"}.}

\item{logger}{\code{Character} scalar, seismic logger name. Must be 
present in the logger library (\code{list_logger}) or parameters must be
added manually. Default is \code{"Cube"}.}

\item{p}{\code{Numeric} scalar, proportion of signal to be tapered. Default
is\code{10^-6}.}

\item{waterlevel}{\code{Numeric} scalar, waterlevel value for frequency
division, default is \code{10^-6}.}
}
\value{
\code{Numeric} scalar, deconvolved signal.
}
\description{
The function removes the instrument response from a signal vector.
}
\examples{

## load example data set
data(rockfall)

## deconvolve signal
rockfall_decon <- signal_deconvolve(data = rockfall,
                                    dt = 1/200, 
                                    sensor = "TC120s", 
                                    logger = "Cube3ext")

## filter signal
rockfall_filter <- signal_filter(data = rockfall_decon, 
                                 dt = 1/200, 
                                 f = c(1,90))

## plot both signals
par_original <- par(no.readonly = TRUE)
par(mfcol = c(2, 1))

plot(x = t, 
     y = rockfall, 
     main = "Rockfall, raw signal", 
     ylab = "Counts", 
     type = "l")

plot(x = t, 
     y = rockfall_filter, 
     main = "Rockfall, deconvolved and filtered signal", 
     ylab = "m/s", 
     type = "l")
par(par_original)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}

