% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_deconvolve.R
\name{signal_deconvolve}
\alias{signal_deconvolve}
\title{Deconvolve a signal vector.}
\usage{
signal_deconvolve(data, dt, sensor = "TC120s", logger = "Cube3ext",
  p = 10^-6, waterlevel = 10^-6)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors, input signal vector.}

\item{dt}{\code{Numeric} scalar, sampling rate.}

\item{sensor}{\code{Character} scalar or \code{list} object, 
seismic sensor name. Must be present in the sensor library 
(\code{list_sensor}) or parameters must be added manually (see 
examples). Default is \code{"TC120s"}.}

\item{logger}{\code{Character} scalar, seismic logger name. Must be 
present in the logger library (\code{list_logger}) or parameters must be
added manually. Default is \code{"Cube"}.}

\item{p}{\code{Numeric} scalar, proportion of signal to be tapered. Default
is\code{10^-6}.}

\item{waterlevel}{\code{Numeric} scalar, waterlevel value for frequency
division, default is \code{10^-6}.}
}
\value{
\code{Numeric} vector or list of vectors, deconvolved signal.
}
\description{
The function removes the instrument response from a signal vector.
}
\examples{

## load example data set
data(rockfall)

## deconvolve signal
rockfall_decon <- signal_deconvolve(data = rockfall,
                                    dt = 1/200, 
                                    sensor = "TC120s", 
                                    logger = "Cube3ext")

## plot time series
plot(x = t, 
     y = rockfall_decon, 
     main = "Rockfall, deconvolved signal", 
     ylab = "m/s", 
     type = "l")
 
## add new logger manually
logger_new <- list_logger()[[1]]

## add logger data
logger_new$ID <- "logger_new"
logger_new$name <- "logger_new"
logger_new$AD <- 2.4414e-07

## deconvolve signal with new logger
rockfall_decon <- signal_deconvolve(data = rockfall,
                                    dt = 1/200, 
                                    sensor = "TC120s", 
                                    logger = logger_new)                     
}
\author{
Michael Dietze
}
\keyword{eseis}

