% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_demean.R
\name{signal_demean}
\alias{signal_demean}
\title{Remove mean of signal vector.}
\usage{
signal_demean(data)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors, input signal vector.}
}
\value{
\code{Numeric} vector or list of vectors, data set with mean 
subtracted.
}
\description{
The function removes the mean from a signal vector.
}
\examples{

## load example data set
data(rockfall)

## remove mean from data set
rockfall_demean <- signal_demean(data = rockfall)

## compare data ranges
range(rockfall)
range(rockfall_demean)

## chow mean of initial signal
mean(rockfall)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
