% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_hilbert.R
\name{signal_hilbert}
\alias{signal_hilbert}
\title{Calculate Hilbert transform.}
\usage{
signal_hilbert(data)
}
\arguments{
\item{data}{\code{eseis} object, \code{numeric} vector or list of 
objects, data set to be processed.}
}
\value{
\code{Numeric} vector or list of vectors, Hilbert transform.
}
\description{
The function calculates the Hilbert transform of the input signal vector.
}
\examples{

## load example data
data(rockfall)

## calculate hilbert transform
rockfall_h <- signal_hilbert(data = rockfall_eseis)

}
\author{
Michael Dietze
}
\keyword{eseis}
