% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_padd.R
\name{signal_padd}
\alias{signal_padd}
\title{Padd signal with zeros.}
\usage{
signal_padd(data)
}
\arguments{
\item{data}{\code{Numeric} vector or list of vectors, input signal vector}
}
\value{
\code{Numeric} vector or list of vectors, signal vector with 
added zeros.
}
\description{
The function adds zeros to the input vector to reach a length,
corresponding to the next higher power of two.
}
\examples{

## load example data set
data(rockfall)

## padd with zeros
rockfall_padd <- signal_padd(data = rockfall_z)

## compare lengths
length(rockfall_z)
length(rockfall_padd)
                     
}
\author{
Michael Dietze
}
\keyword{eseis}
