% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseis-package.R
\docType{data}
\name{t}
\alias{t}
\title{Time vector of a rockfall event.}
\format{The format is: POSIXct[1:98400], format: "2015-04-06 13:16:54" ...}
\usage{
t
}
\description{
The dataset comprises the time vector corresponding the to seismic signal
of the rockfall event from the example data set "rockfall".
}
\examples{

## load example data set
data(rockfall)

}
\keyword{datasets}
